requirements.txt
pandas
matplotlib
seaborn
geolite2
flask
python-dotenv
streamlit
